requirements.txt
python-dotenv
requests
tqdm
beautifulsoup4
langchain
langchain-openai
langchain-pinecone
langchain-core
langchain-community
langgraph
pinecone-client
streamlit